Converts the docket entry number to a bigint, because CA3 seems to use docket numbers in the BILLIONS.

+++ cl/search/models.py
@@ -302,7 +302,7 @@ class DocketEntry(models.Model):
         null=True,
         blank=True,
     )
-    entry_number = models.PositiveIntegerField(
+    entry_number = models.BigIntegerField(
         help_text="# on the PACER docket page.",
     )
     description = models.TextField(
